[Tests] Separated out testing from one file to standard separation of concerns.
